+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog mip.rs
 
-## Next
+## v0.2.3 - 28 May 2026
+
+This is the last version with macOS and Windows support. Future versions will
+be Linux-only, using GTK4 and WebKitGTK 6.0.
+
+- temp files are now written to system temp directory instead of next to the markdown file (#2, #11)
+- temp directory is cleaned up on window close
+- update nixpkgs flake to fix webkitgtk EGL crash on wayland
 - add flake.nix
 - get it working again after cargo update - 2 april 2025